refactor(#38): 경매 좋아요 long으로 변경
